Normalize player names with configurable case handling and centralized processing.
Extracts player name processing logic into a shared utils module to ensure consistent handling across worksheets. Adds configurable name case transformation (UPPER, LOWER, TITLE) and makes the name processing function public for reuse across modules.
